Complete reworking of tests, part 1
